+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/filter/controller/FilterController.java
@@ -1,7 +1,7 @@
 package codesquad.kr.gyeonggidoidle.issuetracker.domain.filter.controller;
 
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.filter.controller.response.FilterListResponse;
-import codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.service.IssueService;
+import codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.service.FilterService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -12,15 +12,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("api/filters")
 public class FilterController {
 
-    private final IssueService issueService;
+    private final FilterService filterService;
 
     @GetMapping("/main")
     public FilterListResponse readFilters() {
-        return FilterListResponse.from(issueService.readFilters());
+        return FilterListResponse.from(filterService.readFilters());
     }
 
     @GetMapping("/detail")
     public FilterListResponse readFiltersFromIssue() {
-        return FilterListResponse.from(issueService.readFiltersFromIssue());
+        return FilterListResponse.from(filterService.readFiltersFromIssue());
     }
 }
