+++ be/src/test/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/issue/controller/IssueControllerTest.java
@@ -35,7 +35,7 @@ class IssueControllerTest {
     @Test
     void readOpenIssues() throws Exception {
         //given
-        given(issueService.readFilteredIssues(anyString())).willReturn(createDummyFilterInformation());
+        given(issueService.read(anyString())).willReturn(createDummyFilterInformation());
 
         //when
         ResultActions resultActions = mockMvc.perform(get("/api/issues?q=is%3Aopen"));
@@ -56,7 +56,7 @@ class IssueControllerTest {
     @Test
     void readClosedIssues() throws Exception {
         //given
-        given(issueService.readFilteredIssues(anyString())).willReturn(createDummyFilterInformation());
+        given(issueService.read(anyString())).willReturn(createDummyFilterInformation());
 
         //when
         ResultActions resultActions = mockMvc.perform(get("/api/issues?q=is%3Aclosed"));
